umnyj-yurist/requirements.txt
# ============================================
# ЗАВИСИМОСТИ УМНЫЙ ЮРИСТ AI
# Версия: 2.0
# ============================================

# Flask - веб-фреймворк для создания сервера
flask==3.0.0

# Python-dotenv - загрузка переменных окружения из .env файла
python-dotenv==1.0.0

# OpenAI - клиент для работы с OpenRouter API
openai==1.12.0

# Werkzeug - утилиты для безопасности (secure_filename)
werkzeug==3.0.1

# Pillow - работа с изображениями (если понадобится OCR)
Pillow==10.2.0

# Requests - HTTP-запросы
requests==2.31.0